Remove extra call findRoot()
