+++ study-algorithm/rbtree/rbtree/main.go
@@ -128,20 +128,6 @@ func rotate(n *Node) *Node {
 	}
 }
 
-type placeType int
-
-const (
-	root placeType = 0 + iota
-	left
-	right
-)
-
-type place struct {
-	t      placeType
-	tree   *RBTree
-	parent *Node
-}
-
 func (tree *RBTree) Delete(k Key) (Value, error) {
 	p, cur := find(tree.root, k)
 	if cur == nil {
@@ -207,53 +193,6 @@ func (tree *RBTree) place(p, n *Node) place {
 	} else {
 		return place{t: right, tree: nil, parent: p}
 	}
-}
-
-func (p place) Node() *Node {
-	switch p.t {
-	case root:
-		return p.tree.root
-	case left:
-		return p.parent.l
-	case right:
-		return p.parent.r
-	}
-	return nil
-}
-
-func (p place) setOnPlace(n *Node) {
-	var link **Node
-	switch p.t {
-	case root:
-		link = &(p.tree.root)
-	case left:
-		link = &(p.parent.l)
-	case right:
-		link = &(p.parent.r)
-	}
-
-	if n == nil {
-		(*link) = nil
-		return
-	} else if (*link) == nil {
-		*link = n
-	} else {
-		(*link).copyFrom(n)
-	}
-
-	if (*link).l != nil {
-		(*link).l.p = (*link)
-	}
-	if (*link).r != nil {
-		(*link).r.p = (*link)
-	}
-
-	if p.t == root {
-		(*link).p = (*link)
-		return
-	}
-	(*link).p = p.parent
-	return
 }
 
 func findSubstitue(n *Node) (p, cur *Node) {
@@ -322,7 +261,6 @@ func (tree *RBTree) recoverRankLeft(p *Node) {
 		tree.recoverRankLeft(new.l)
 		return
 	}
-	tree.root = findRoot(p)
 	return
 }
 
@@ -357,6 +295,5 @@ func (tree *RBTree) recoverRankRight(p *Node) {
 		tree.recoverRankRight(new.r)
 		return
 	}
-	tree.root = findRoot(p)
 	return
 }
